+++ crates/core/src/pattern/not.rs
@@ -23,49 +23,6 @@ impl Not {
     pub fn new(pattern: Pattern) -> Self {
         Self { pattern }
     }
-
-    pub(crate) fn from_node(
-        node: &Node,
-        context: &CompilationContext,
-        vars: &mut BTreeMap<String, usize>,
-        vars_array: &mut Vec<Vec<VariableSourceLocations>>,
-        scope_index: usize,
-        global_vars: &mut BTreeMap<String, usize>,
-        logs: &mut AnalysisLogs,
-    ) -> Result<Self> {
-        let pattern = node
-            .child_by_field_name("pattern")
-            .ok_or_else(|| anyhow!("missing pattern of patternNot"))?;
-        let range: Range = pattern.range().into();
-        let pattern = Pattern::from_node(
-            &pattern,
-            context,
-            vars,
-            vars_array,
-            scope_index,
-            global_vars,
-            false,
-            logs,
-        )?;
-        if pattern.iter().any(|p| {
-            matches!(
-                p,
-                PatternOrPredicate::Pattern(Pattern::Rewrite(_))
-                    | PatternOrPredicate::Predicate(Predicate::Rewrite(_))
-            )
-        }) {
-            let log = AnalysisLogBuilder::default()
-                .level(441_u16)
-                .file(context.file)
-                .source(context.src)
-                .position(range.start)
-                .range(range)
-                .message("Warning: rewrites inside of a not will never be applied")
-                .build()?;
-            logs.push(log);
-        }
-        Ok(Self::new(pattern))
-    }
 }
 
 impl Name for Not {
