+++ crates/core/src/pattern/patterns.rs
@@ -63,7 +63,8 @@ use crate::{
         includes_compiler::IncludesCompiler, like_compiler::LikeCompiler,
         limit_compiler::LimitCompiler, list_index_compiler::ListIndexCompiler,
         log_compiler::LogCompiler, maybe_compiler::MaybeCompiler, modulo_compiler::ModuloCompiler,
-        multiply_compiler::MultiplyCompiler, CompilationContext, NodeCompiler,
+        multiply_compiler::MultiplyCompiler, not_compiler::NotCompiler, CompilationContext,
+        NodeCompiler,
     },
 };
 use anyhow::{anyhow, bail, Result};
@@ -759,7 +760,7 @@ impl Pattern {
                 global_vars,
                 logs,
             )?))),
-            "patternNot" => Ok(Pattern::Not(Box::new(Not::from_node(
+            "patternNot" => Ok(Pattern::Not(Box::new(NotCompiler::from_node(
                 node,
                 context,
                 vars,
