+++ crates/core/src/pattern_compiler/mod.rs
@@ -26,6 +26,7 @@ pub(crate) mod maybe_compiler;
 pub(crate) mod modulo_compiler;
 pub(crate) mod multiply_compiler;
 mod node_compiler;
+pub(crate) mod not_compiler;
 pub(crate) mod step_compiler;
 
 pub(crate) use compiler::{parse_one, CompilationContext};
